Move rule to Hotspots directory
